many things, most are only half-finished

+++ shaders/include/utility/sequence.glsl
@@ -15,17 +15,17 @@ const float phi1 = 1.61803398874989484820458683436563; // = phi
 const float phi2 = 1.32471795724474602596090885447809; // = plastic constant, plastic ratio, etc
 const float phi3 = 1.220744084605759475361685349108831;
 
-float R1(int n) {
+float R1(float n) {
 	const float s0 = 0.5;
 	const float alpha = 1.0 / phi1;
 	return fract(s0 + n * alpha);
 }
-vec2 R2(int n) {
+vec2 R2(float n) {
 	const float s0 = 0.5;
 	const vec2 alpha = 1.0 / vec2(phi2, phi2 * phi2);
 	return fract(s0 + n * alpha);
 }
-vec3 R3(int n) {
+vec3 R3(float n) {
 	const float s0 = 0.5;
 	const vec3 alpha = 1.0 / vec3(phi3, phi3 * phi3, phi3 * phi3 * phi3);
 	return fract(s0 + n * alpha);
